code: move reducer out into separate module
